+++ src/BaseFileDecoder.cs
@@ -22,7 +22,7 @@ public class BaseFileDecoder {
 		filePath = _filePath;
 
 		File f = new File();
-		Error e = f.Open(filePath, (int)File.ModeFlags.Read);
+		Error e = f.Open(filePath, File.ModeFlags.Read);
 
 		if (e != Error.Ok) {
 			throw new ArgumentException("Error opening file: " + filePath + " - Error " + e.ToString());
@@ -44,7 +44,7 @@ public class BaseFileDecoder {
 
 		switch (method) {
 			case (DecodingMethod.None):
-				data = fileIn.GetBuffer(fileIn.GetLen());
+				data = fileIn.GetBuffer((int)fileIn.GetLen());
 				break;
 			case (DecodingMethod.xtRLE):
 				data = DecodeXTRLE(fileIn);
@@ -77,7 +77,7 @@ public class BaseFileDecoder {
 		byte[] data;
 
 		fileIn.Seek(10); //Skip header
-		int finalSize = fileIn.Get32(); //Get file size
+		int finalSize = (int)fileIn.Get32(); //Get file size
 
 		data = null;
